requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.1.0
Flask==3.0.3
frozendict==2.4.4
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.7
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multitasking==0.0.11
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
platformdirs==4.2.2
plotly==5.24.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.2.0
redis==5.0.8
requests==2.31.0
selenium==4.24.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
wsproto==1.2.0
yfinance==0.2.43



